+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 0.1.3
 
+* Added `EventStoreQueryRange` to the APIs of the event store and the globally
+  ordered event store. This allows the user to specify optional start and stop
+  points for the query.
 * Added versions of the `MonadState` event stores that can include other state
   along with the `EventMap`. These are called `embeddedStateEventStore` and
   `embeddedStateGloballyOrderedEventStore`.